+++ src/pharmpy/internals/fs/symlink.py
@@ -8,6 +8,12 @@ def create_directory_symlink(link_path: Path, target_path: Path):
     # Creating a symlink in Windows requires elevated privileges, but creating a junction
     # doesn't. The junction is mostly equivalent to symlinks so should be fine.
     if os.name == 'nt':
-        subprocess.check_call(f'mklink /J "{link_path}" "{target_path}" >nul 2>&1', shell=True)
+        # DEVNULL needed to avoid WinError 6: The handle is invalid
+        subprocess.check_call(
+            f'mklink /J "{link_path}" "{target_path}" >nul 2>&1',
+            shell=True,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        )
     else:
         link_path.symlink_to(target_path)